timeLayout
The row with gallery, capture and flip buttons used to shift up when recording was in progress. this line is just to take up the free space and avoid the shift.
It could be useful to use it to say the length of the recording (reason for the name) but that's just a detail.